app/requirements.txt
anyio==3.7.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
kafka-python==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.6
pydantic==1.10.9
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
